+++ BLCMM/src/blcmm/gui/panels/AboutPanel.java
@@ -74,6 +74,7 @@ public final class AboutPanel extends JPanel {
                 + "  - Xerial's sqlite-jdbc, available under the Apache License v2.0<br/>"
                 + "  - Some icons from Dave Gandy's Font Awesome set, available under CC BY 3.0<br/>"
                 + "  - An icon from Fathema Khanom's User Interface set, available under Flaticon's Free License<br/>"
+                + "  - An icon from Smashicons' Essential Collection set, available under Flaticon's Free License<br/>"
         );
         int wLog = 64, wPay = 100;
         mainLabel.setIcon(new ImageIcon(IconManager.getBLCMMIcon(wLog)));